+++ MoPubSDK/Internal/Utility/MPGlobal.h
@@ -73,7 +73,7 @@ typedef NSUInteger MPInterstitialOrientationType;
 /*
  * Returns string with reserved/unsafe characters encoded.
  */
-- (NSString *)URLEncodedString;
+- (NSString *)mp_URLEncodedString;
 
 @end
 
@@ -81,7 +81,7 @@ typedef NSUInteger MPInterstitialOrientationType;
 
 @interface UIDevice (MPAdditions)
 
-- (NSString *)hardwareDeviceName;
+- (NSString *)mp_hardwareDeviceName;
 
 @end