+++ MoPubSDK/Internal/Utility/MPGlobal.m
@@ -204,7 +204,7 @@ NSArray *MPConvertStringArrayToURLArray(NSArray *strArray)
 
 @implementation NSString (MPAdditions)
 
-- (NSString *)URLEncodedString
+- (NSString *)mp_URLEncodedString
 {
     NSString *result = (NSString *)CFBridgingRelease(CFURLCreateStringByAddingPercentEscapes(NULL,
                                                                            (CFStringRef)self,
@@ -220,7 +220,7 @@ NSArray *MPConvertStringArrayToURLArray(NSArray *strArray)
 
 @implementation UIDevice (MPAdditions)
 
-- (NSString *)hardwareDeviceName
+- (NSString *)mp_hardwareDeviceName
 {
     size_t size;
     sysctlbyname("hw.machine", NULL, &size, NULL, 0);