+++ MoPubSampleApp/Views/MPNativeVideoTableViewAdPlacerView.h
@@ -1,20 +1,16 @@
 //
-//  MPStaticNativeAdView.h
-//  MoPubSampleApp
-//
-//  Copyright (c) 2015 MoPub. All rights reserved.
-//
+//  MPNativeVideoTableViewAdPlacerView.h
 
-#import <UIKit/UIKit.h>
 #import "MPNativeAdRendering.h"
 
-@interface MPStaticNativeAdView : UIView <MPNativeAdRendering>
+@interface MPNativeVideoTableViewAdPlacerView : UIView <MPNativeAdRendering>
 
 @property (strong, nonatomic) UILabel *titleLabel;
 @property (strong, nonatomic) UILabel *mainTextLabel;
 @property (strong, nonatomic) UIImageView *iconImageView;
 @property (strong, nonatomic) UIImageView *mainImageView;
-@property (strong, nonatomic) UIImageView *privacyInformationIconImageView;
+@property (strong, nonatomic) UIView *videoView;
+@property (strong, nonatomic) UIImageView *DAAIconImageView;
 @property (strong, nonatomic) UILabel *ctaLabel;
 
 @end